requirements.txt
selenium
webdriver-manager
pandas
mysql-connector-python
